Atualização Automática via bash shell

+++ compilar_vps.sh
@@ -160,13 +160,16 @@ VAGRANT_VAGRANTFILE=Vagrantfile.VPS_DEV vagrant ssh<<EOF
 cd /vagrant
 make iso
 
-echo "Download Vagrant & Instalar"
-wget -nv https://releases.hashicorp.com/vagrant/2.2.9/vagrant_2.2.9_x86_64.deb
-sudo dpkg -i vagrant_2.2.9_x86_64.deb
-rm vagrant_2.2.9_x86_64.deb
+if ! command -v vagrant &> /dev/null
+then
+	echo "Download Vagrant & Instalar"
+	wget -nv https://releases.hashicorp.com/vagrant/2.2.9/vagrant_2.2.9_x86_64.deb
+	sudo dpkg -i vagrant_2.2.9_x86_64.deb
+	rm vagrant_2.2.9_x86_64.deb
 
-echo "Instalar plugins do Vagrant"
-vagrant plugin install vagrant-libvirt
+	echo "Instalar plugins do Vagrant"
+	vagrant plugin install vagrant-libvirt
+fi
 
 cd .. 
 EOF
